LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/basic-expected.txt

abc
def

FAIL Minimal scroll anchoring example. assert_equals: expected 250 but got 150
